Convert AddFiresToHitbox
