+++ src/main/java/net/countercraft/movecraft/combat/features/AddFiresToHitbox.java
@@ -1,0 +1,51 @@
+package net.countercraft.movecraft.combat.features;
+
+import net.countercraft.movecraft.craft.Craft;
+import net.countercraft.movecraft.craft.CraftManager;
+import net.countercraft.movecraft.util.MathUtils;
+import net.countercraft.movecraft.util.hitboxes.MutableHitBox;
+import org.bukkit.Location;
+import org.bukkit.configuration.file.FileConfiguration;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockIgniteEvent;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public class AddFiresToHitbox implements Listener {
+    public static boolean AddFiresToHitbox = true;
+
+    public static void load(@NotNull FileConfiguration config) {
+        AddFiresToHitbox = config.getBoolean("AddFiresToHitbox", true);
+    }
+
+    @EventHandler(priority = EventPriority.LOW, ignoreCancelled = true)
+    public void onBlockIgnite(@NotNull BlockIgniteEvent e) {
+        Craft craft = adjacentCraft(e.getBlock().getLocation());
+        if(craft == null || craft.getHitBox().isEmpty())
+            return;
+        if(!(craft.getHitBox() instanceof MutableHitBox))
+            return;
+
+        MutableHitBox hitbox = (MutableHitBox) craft.getHitBox();
+        hitbox.add(MathUtils.bukkit2MovecraftLoc(e.getBlock().getLocation()));
+    }
+
+    @Nullable
+    private Craft adjacentCraft(@NotNull Location location) {
+        Craft craft = CraftManager.getInstance().fastNearestCraftToLoc(location);
+        if(craft == null)
+            return null; //return null if no craft found
+
+        if(MathUtils.locationInHitBox(craft.getHitBox(), location.add(1,0,0))
+                || MathUtils.locationInHitBox(craft.getHitBox(), location.add(-1,0,0))
+                || MathUtils.locationInHitBox(craft.getHitBox(), location.add(0,1,0))
+                || MathUtils.locationInHitBox(craft.getHitBox(), location.add(0,-1,0))
+                || MathUtils.locationInHitBox(craft.getHitBox(), location.add(0,0,1))
+                || MathUtils.locationInHitBox(craft.getHitBox(), location.add(0,0,-1)))
+            return craft;
+
+        return null;
+    }
+}
